Add test project
