+++ __test_project__/package.json
@@ -1,16 +1,18 @@
 {
-  "name": "test",
+  "name": "test_project",
   "version": "1.0.0",
+  "main": "index.js",
   "license": "MIT",
+  "scripts": {
+    "build": "webpack"
+  },
   "devDependencies": {
     "css-loader": "^2.1.0",
     "mini-css-extract-plugin": "^0.5.0",
     "postcss": "^7.0.14",
     "postcss-assets-webpack-plugin": "^3.0.0",
-    "postcss-extend-rule": "^2.0.0",
     "postcss-loader": "^3.0.0",
-    "postcss-modules-extend-class": "0.0.1",
-    "postcss-modules-extend-from": "0.0.1",
+    "postcss-modules-extend-rule": "../",
     "webpack": "^4.29.2",
     "webpack-cli": "^3.2.3"
   },
