+++ __test_project__/src/border.css
@@ -1,0 +1,3 @@
+.border {
+  border: 1px solid black;
+}
