+++ __test_project__/src/colors.css
@@ -1,0 +1,15 @@
+.blueForeground {
+  color: blue;
+}
+
+.blueBackground {
+  background: blue;
+}
+
+.redForeground {
+  color: red;
+}
+
+.redBackground {
+  background: red;
+}
