--- test/src/blueForeground.css
@@ -1,3 +1,0 @@
-.blue-foreground {
-  color: blue;
-}
