--- test/src/index.css
@@ -1,8 +1,0 @@
-.red-background {
-  background: red;
-}
-
-.blueOnRed {
-  @extend .red-background;
-  @extend .blue-foreground from './blueForeground.css';
-}
